fix: Rename cssUtils since otherwise css.js can be thought to be a CSS files
The original root is a badly configured bundler
